Fixed the auto lighting config on spawn
