+++ fbw-a32nx/src/base/flybywire-aircraft-a320-neo/html_ui/Pages/A32NX_Core/A32NX_Core.js
@@ -70,51 +70,8 @@ class A32NX_Core {
                 moduleDefinition.module.init();
             }
         });
-        this.initLighting();
 
         this.isInit = true;
-    }
-
-    initLighting() {
-        /** automatic brightness based on ambient light, [0, 1] scale */
-        const autoBrightness = Math.max(15, Math.min(85, SimVar.GetSimVarValue('GLASSCOCKPIT AUTOMATIC BRIGHTNESS', 'percent')));
-
-        // DOME
-        this.setPotentiometer(7, 0);
-        // MAIN FLOOD
-        this.setPotentiometer(83, autoBrightness < 50 ? 20 : 0);
-        // FCU INTEG
-        this.setPotentiometer(84, autoBrightness < 50 ? 1.5 * autoBrightness : 0);
-        // MAIN & PED INTEG
-        this.setPotentiometer(85, autoBrightness < 50 ? 1.5 * autoBrightness : 0);
-        // OVHD INTEG
-        this.setPotentiometer(86, autoBrightness < 50 ? 1.5 * autoBrightness : 0);
-        // FCU Displays
-        this.setPotentiometer(87, autoBrightness);
-        // CAPT PFD DU
-        this.setPotentiometer(88, autoBrightness);
-        // CAPT ND DU
-        this.setPotentiometer(89, autoBrightness);
-        // F/O PFD DU
-        this.setPotentiometer(90, autoBrightness);
-        // F/O ND DU
-        this.setPotentiometer(91, autoBrightness);
-        // Upper ECAM DU
-        this.setPotentiometer(92, autoBrightness);
-        // Lower ECAM DU
-        this.setPotentiometer(93, autoBrightness);
-        // CAPT MCDU
-        SimVar.SetSimVarValue('L:A32NX_MCDU_L_BRIGHTNESS', 'number', 8 * autoBrightness / 100);
-        // FO MCDU
-        SimVar.SetSimVarValue('L:A32NX_MCDU_R_BRIGHTNESS', 'number', 8 * autoBrightness / 100);
-        // CAPT DCDU
-        SimVar.SetSimVarValue('L:A32NX_PANEL_DCDU_L_BRIGHTNESS', 'number', autoBrightness / 100);
-        // FO DCDU
-        SimVar.SetSimVarValue('L:A32NX_PANEL_DCDU_R_BRIGHTNESS', 'number', autoBrightness / 100);
-    }
-
-    setPotentiometer(potentiometer, brightness) {
-        Coherent.call('TRIGGER_KEY_EVENT', 'LIGHT_POTENTIOMETER_SET', false, potentiometer, brightness, 0, 0);
     }
 
     update() {
